test(core): Refactor TicketBase model API Fields render test Suite to PyTest
ref: #892 #889
