+++ app/core/tests/functional/ticket_base/conftest.py
@@ -22,3 +22,11 @@ def create_serializer():
 
 
     yield ModelSerializer
+
+
+@pytest.fixture( scope = 'class', autouse = True)
+def model_kwargs(request, kwargs_ticketbase):
+
+    request.cls.kwargs_create_item = kwargs_ticketbase.copy()
+
+    yield kwargs_ticketbase.copy()
